requirements.txt
torch>=2.0.0
torchvision>=0.15.0
gym-super-mario-bros>=7.4.0
nes-py>=8.2.1
numpy>=1.24.0
opencv-python>=4.8.0
PyYAML>=6.0
tensorboard>=2.14.0
nicegui>=1.4.0

